lexer: move `Base` type into lexer codebase
