+++ compiler/hash-token/src/lib.rs
@@ -3,8 +3,6 @@
 pub mod delimiter;
 pub mod keyword;
 
-use std::fmt::Display;
-
 use delimiter::{Delimiter, DelimiterVariant};
 use hash_source::{
     constant::{InternedFloat, InternedInt, InternedStr},
@@ -274,46 +272,6 @@ impl std::fmt::Display for TokenKind {
                 write!(f, "{}", String::from(*ident))
             }
             TokenKind::Err => write!(f, "<error>"),
-        }
-    }
-}
-
-/// Represents the featured base for numeric literals.
-#[derive(Debug)]
-pub enum Base {
-    /// Binary base, denoted in literals as `0b101010`
-    Binary,
-    /// Octal base, denoted in literals as `0o26317261`
-    Octal,
-    /// Decimal base, written as `102391`
-    Decimal,
-    /// Hexadecimal base, written as `0xdeadbeef`
-    Hex,
-    /// Unsupported base, the language doesn't support the
-    /// provided radix as a base.
-    Unsupported,
-}
-
-impl Display for Base {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Base::Binary => write!(f, "binary"),
-            Base::Octal => write!(f, "octal"),
-            Base::Decimal => write!(f, "decimal"),
-            Base::Hex => write!(f, "hexadecimal"),
-            Base::Unsupported => write!(f, "unsupported base"),
-        }
-    }
-}
-
-impl From<u32> for Base {
-    fn from(radix: u32) -> Self {
-        match radix {
-            2 => Base::Binary,
-            8 => Base::Octal,
-            10 => Base::Decimal,
-            16 => Base::Hex,
-            _ => Base::Unsupported,
         }
     }
 }
